+++ src/collectors/BranchesCollector.ts
@@ -21,7 +21,6 @@ export class BranchesCollector {
       this.contentRepositoryBrowser.listBranches(ownerAndRepoName.owner, ownerAndRepoName.repoName),
       this.gitInspector.getStatus(),
     ]);
-    // TODO: test gitlab
     const defaultBranch = branches.items.find((branch) => branch.type === 'default');
 
     return { current: status.current || 'unknown', default: defaultBranch?.name || 'unknown' };