+++ src/practices/JavaScript/JsGitignoreCorrectlySetPractice.ts
@@ -1,9 +1,7 @@
-import { PracticeEvaluationResult, PracticeImpact, ProgrammingLanguage } from '../../model';
+import { PracticeImpact, ProgrammingLanguage } from '../../model';
 import { DxPractice } from '../DxPracticeDecorator';
-import { PracticeContext } from '../../contexts/practice/PracticeContext';
-import { PracticeBase } from '../PracticeBase';
 import { ReportDetailType } from '../../reporters/ReporterData';
-import { FixerContext } from '../../contexts/fixer/FixerContext';
+import { GitignoreCorrectlySetPracticeBase } from '../common/GitignoreCorrectlySetPracticeBase';
 
 @DxPractice({
   id: 'JavaScript.GitignoreCorrectlySet',
@@ -14,64 +12,20 @@ import { FixerContext } from '../../contexts/fixer/FixerContext';
   url: 'https://github.com/github/gitignore/blob/master/Node.gitignore',
   dependsOn: { practicing: ['LanguageIndependent.GitignoreIsPresent'] },
 })
-export class JsGitignoreCorrectlySetPractice extends PracticeBase {
-  private parsedGitignore: string[] = [];
-
-  async isApplicable(ctx: PracticeContext): Promise<boolean> {
-    return ctx.projectComponent.language === ProgrammingLanguage.JavaScript;
+export class JsGitignoreCorrectlySetPractice extends GitignoreCorrectlySetPracticeBase {
+  constructor() {
+    super();
+    this.applicableLanguages = [ProgrammingLanguage.JavaScript];
+    this.ruleChecks = [
+      // node_modules
+      { regex: /node_modules/, fix: '/node_modules' },
+      // misc
+      { regex: /coverage/, fix: '/coverage' },
+      { regex: /\.log/, fix: '*.log' },
+    ];
   }
 
-  async evaluate(ctx: PracticeContext): Promise<PracticeEvaluationResult> {
-    if (!ctx.root.fileInspector) {
-      return PracticeEvaluationResult.unknown;
-    }
-
-    const parseGitignore = (gitignoreFile: string) => {
-      return gitignoreFile
-        .toString()
-        .split(/\r?\n/)
-        .filter((content) => content.trim() !== '' && !content.startsWith('#'));
-    };
-    const content = await ctx.root.fileInspector.readFile('.gitignore');
-    const parsedGitignore = parseGitignore(content);
-    this.parsedGitignore = parsedGitignore;
-
-    // node_modules
-    const nodeModulesRegex = parsedGitignore.find((value: string) => /node_modules/.test(value));
-    // misc
-    const coverageRegex = parsedGitignore.find((value: string) => /coverage/.test(value));
-    const errorLogRegex = parsedGitignore.find((value: string) => /\.log/.test(value));
-
-    if (nodeModulesRegex && errorLogRegex && coverageRegex) {
-      return PracticeEvaluationResult.practicing;
-    }
-
-    this.setData();
-    return PracticeEvaluationResult.notPracticing;
-  }
-
-  async fix(ctx: FixerContext) {
-    const inspector = ctx.fileInspector?.basePath ? ctx.fileInspector : ctx.root.fileInspector;
-    if (!inspector) return;
-    // node_modules
-    const nodeModulesRegex = this.parsedGitignore.find((value: string) => /node_modules/.test(value));
-    // misc
-    const coverageRegex = this.parsedGitignore.find((value: string) => /coverage/.test(value));
-    const errorLogRegex = this.parsedGitignore.find((value: string) => /\.log/.test(value));
-    const fixes = [
-      nodeModulesRegex ? undefined : '/node_modules',
-      coverageRegex ? undefined : '/coverage',
-      errorLogRegex ? undefined : '*.log',
-    ]
-      .filter(Boolean)
-      .concat(''); // append newline if we add something
-
-    if (fixes.length > 1) fixes.unshift(''); // if there is something to add, make sure we start with newline
-
-    await inspector.appendFile('.gitignore', fixes.join('\n'));
-  }
-
-  private setData() {
+  protected setData() {
     this.data.details = [
       {
         type: ReportDetailType.text,